requirements.txt
streamlit
python-dotenv
google-generativeai
pandas
scikit-learn
matplotlib
seaborn
